pyproject.toml


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["z007"]

[project]
name = "z007"
version = "0.1.0"
description = "Micro agent with tool support and MCP integration."
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Igor Okulist", email = "okigan@gmail.com" }
]
keywords = ["ai", "llm", "agent", "mcp", "bedrock", "tools", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "anyio>=4.10.0",
    "boto3",
    "botocore",
    "colorlog>=6.9.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/okigan/z007"
"Repository" = "https://github.com/okigan/z007"
"Issues" = "https://github.com/okigan/z007/issues"
"Changelog" = "https://github.com/okigan/z007/releases"

[project.scripts]
z007 = "z007.main:main"

[project.optional-dependencies]
dev = [
    "mypy-boto3-bedrock-runtime>=1.40.0",
    "mypy-boto3-bedrock>=1.40.0",
    "boto3-stubs[essential]>=1.40.0",
    "mypy>=1.17.1",
    "ruff>=0.12.11",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[tool.uv]
dev-dependencies = [
    "mypy-boto3-bedrock-runtime>=1.40.0",
    "mypy-boto3-bedrock>=1.40.0",
    "boto3-stubs[essential]>=1.40.0",
    "mypy>=1.17.1",
    "ruff>=0.12.11",
]

[tool.mypy]
# Basic strict settings we can achieve
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# We can be more strict now with better type stubs
disallow_untyped_defs = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_subclassing_any = false

# Third-party libraries without proper stubs
[[tool.mypy.overrides]]
module = "colorlog"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "anyio.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "select"
ignore_missing_imports = true
